requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
blinker==1.9.0
bottle==0.13.2
cffi==1.17.1
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.56.0
indeterminatebeam==2.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
narwhals==1.32.0
numpy==2.2.4
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
plotly==6.0.1
proxy_tools==0.1.0
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pyparsing==3.2.3
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6-WebEngine==6.8.0
PyQt6-WebEngine-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pywebview==5.4
pywin32-ctypes==0.2.3
QtPy==2.4.3
setuptools==78.1.0
six==1.17.0
sympy==1.13.3
typing-inspection==0.4.0
typing_extensions==4.13.0
waitress==3.0.2
Werkzeug==3.1.3
